+++ packages/hel-micro-react/src/consts/ver.ts
@@ -1,1 +1,1 @@
-export const VER = '4.9.0';
+export const VER = '4.9.1';