+++ packages/hel-micro/src/index.ts
@@ -15,10 +15,11 @@ import type {
 } from 'hel-micro-core';
 import * as apis from './apis';
 import * as ins from './ins';
+export type { ApiMode, IEmitAppInfo, ISubApp, ISubAppVersion, Platform } from 'hel-types';
 export type { IGroupedStyleList, IPreFetchAppOptions, IPreFetchLibOptions, IPreFetchOptionsBase, IStyleDataResult } from './types';
 
 /** 暂时解决部分 v3 版本用户升级到 v4 后，原来使用 helMicro.core.IAppAndVer 标准类型出错的问题 */
-export namespace core {
+export declare namespace core {
   export type IAppAndVer = t1;
   export type IControlPreFetchOptions = t2;
   export type IHelMeta = t3;